rename/mobile_card/#119 : 모바일카드 도메인명을 변경하였습니다.
