+++ src/main/java/com/ghbt/ghbt_starbucks/api/user_has_mobilecard/service/UserHasMobileCardServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ghbt.ghbt_starbucks.api.user_has_mobilecard.service;
 
-import com.ghbt.ghbt_starbucks.api.mobilecard.model.MobileCard;
-import com.ghbt.ghbt_starbucks.api.mobilecard.repository.MobileCardRepository;
+import com.ghbt.ghbt_starbucks.api.mobile_card.model.MobileCard;
+import com.ghbt.ghbt_starbucks.api.mobile_card.repository.MobileCardRepository;
 import com.ghbt.ghbt_starbucks.api.user.model.User;
 import com.ghbt.ghbt_starbucks.api.user_has_mobilecard.dto.RequestMobileCard;
 import com.ghbt.ghbt_starbucks.api.user_has_mobilecard.dto.ResponseMobileCardAndUser;
@@ -32,7 +32,6 @@ public class UserHasMobileCardServiceImpl implements IUserHasMobileCardService {
         return userHasMobileCards.stream()
             .map(ResponseMobileCardAndUser::from)
             .collect(Collectors.toList());
-
     }
 
     @Override
